move class back to old location with new import
